src/backend/requirements.txt
# Core dependencies
fastapi~=0.115
uvicorn[standard]~=0.32
pydantic~=2.9
python-multipart~=0.0.12

# Azure dependencies
azure-storage-blob~=12.19
azure-identity~=1.15

# OpenTelemetry dependencies
opentelemetry-api~=1.29
opentelemetry-sdk~=1.29
opentelemetry-instrumentation-fastapi~=0.50b0
opentelemetry-exporter-otlp-proto-grpc~=1.29

# Auth dependencies
PyJWT[crypto]~=2.9

# Testing dependencies
pytest~=9.0
pytest-asyncio~=1.3
pytest-cov~=7.0
httpx~=0.28
